Give the mechs shoes
It's a long story, ok? James wanted his mech to have shoes, so I figured I'd just add a button to tack some shoes on that bad boy, but then he said he wanted ONLY his mech to have shoes, and I couldn't just put a button there that everybody could press, where's the fun in that? But I don't have any authentication or server-side, so what to do? Well, I just put a key handler on the display component that listens for a passcode that will display shoes. Is it a lot of work for an inside joke? Kindof. But I had to.