Hoplite: Archer changes, floor 12 flash, shop
The beginnings of a shop! The very beginnings. Not useful yet.
Archers now always try to get line-of-sight, but dont attack the turn after being frightened.
Added a rudimentary flash animation to completing floor 12. It's not ideal, but it gets the job done.
Added the floor 12 portal. It doesn't end the game, currently.